moved schema into individual tables
