+++ src/db/app-schema.ts
@@ -1,0 +1,34 @@
+import {
+  pgTable,
+  integer,
+  varchar,
+  text,
+  timestamp,
+  boolean,
+} from "drizzle-orm/pg-core"
+import { users } from "./auth-schema"
+
+export const projectsTable = pgTable(`projects`, {
+  id: integer().primaryKey().generatedAlwaysAsIdentity(),
+  name: varchar({ length: 255 }).notNull(),
+  description: text(),
+  shared_user_ids: text("shared_user_ids").array().notNull().default([]),
+  created_at: timestamp({ withTimezone: true }).notNull().defaultNow(),
+  owner_id: text("owner_id")
+    .notNull()
+    .references(() => users.id, { onDelete: "cascade" }),
+})
+
+export const todosTable = pgTable(`todos`, {
+  id: integer().primaryKey().generatedAlwaysAsIdentity(),
+  text: varchar({ length: 500 }).notNull(),
+  completed: boolean().notNull().default(false),
+  created_at: timestamp({ withTimezone: true }).notNull().defaultNow(),
+  user_id: text("user_id")
+    .notNull()
+    .references(() => users.id, { onDelete: "cascade" }),
+  project_id: integer("project_id")
+    .notNull()
+    .references(() => projectsTable.id, { onDelete: "cascade" }),
+  user_ids: text("user_ids").array().notNull().default([]),
+})
