+++ src/db/schema.ts
@@ -1,43 +1,10 @@
-import {
-  boolean,
-  integer,
-  pgTable,
-  timestamp,
-  varchar,
-  text,
-} from "drizzle-orm/pg-core"
 import { createSchemaFactory } from "drizzle-zod"
 import { z } from "@hono/zod-openapi"
-export * from "./auth-schema"
 import { users } from "./auth-schema"
+import { projectsTable, todosTable } from "./app-schema"
 
 const { createInsertSchema, createSelectSchema, createUpdateSchema } =
   createSchemaFactory({ zodInstance: z })
-
-export const projectsTable = pgTable(`projects`, {
-  id: integer().primaryKey().generatedAlwaysAsIdentity(),
-  name: varchar({ length: 255 }).notNull(),
-  description: text(),
-  shared_user_ids: text("shared_user_ids").array().notNull().default([]),
-  created_at: timestamp({ withTimezone: true }).notNull().defaultNow(),
-  owner_id: text("owner_id")
-    .notNull()
-    .references(() => users.id, { onDelete: "cascade" }),
-})
-
-export const todosTable = pgTable(`todos`, {
-  id: integer().primaryKey().generatedAlwaysAsIdentity(),
-  text: varchar({ length: 500 }).notNull(),
-  completed: boolean().notNull().default(false),
-  created_at: timestamp({ withTimezone: true }).notNull().defaultNow(),
-  user_id: text("user_id")
-    .notNull()
-    .references(() => users.id, { onDelete: "cascade" }),
-  project_id: integer("project_id")
-    .notNull()
-    .references(() => projectsTable.id, { onDelete: "cascade" }),
-  user_ids: text("user_ids").array().notNull().default([]),
-})
 
 export const selectProjectSchema = createSelectSchema(projectsTable)
 export const createProjectSchema = createInsertSchema(projectsTable)
@@ -61,3 +28,6 @@ export type Todo = z.infer<typeof selectTodoSchema>
 export type UpdateTodo = z.infer<typeof updateTodoSchema>
 
 export const selectUsersSchema = createSelectSchema(users)
+
+export { users, sessions, accounts, verifications } from "./auth-schema"
+export { projectsTable, todosTable } from "./app-schema"
